Fix configDocuments & extract to its own file.
