+++ src/core/store/entities.js
@@ -1,5 +1,5 @@
 import {
-  allPass, anyPass, compose, evolve, ifElse, map, none, pick, prop, reduce,
+  allPass, anyPass, compose, evolve, ifElse, map, prop, reduce,
 } from 'ramda';
 import farm from '../farm';
 import nomenclature from './nomenclature';
@@ -9,7 +9,6 @@ import {
 import { cachingCriteria, evictionCriteria } from './criteria';
 import upsert from '../utils/upsert';
 import parseFilter from '../utils/parseFilter';
-import databases from '../idb/databases';
 import SyncError from './SyncError';
 
 const cacheEntity = (name, criteria, entity) => {
@@ -33,30 +32,6 @@ const flattenSerializedEntity = ({
   id, type, meta, ...attributes, ...flattenRelationships(relationships),
 });
 
-const docsReducer = ([schemata, types], { key, config_entity, json_schema }) => {
-  const _schemata = json_schema ? { ...schemata, [key]: json_schema } : schemata;
-  const _types = config_entity ? { ...types, [key]: config_entity } : types;
-  return [_schemata, _types];
-};
-const partitionConfigDocs = reduce(docsReducer, [{}, {}]);
-const configTypes = ['asset', 'log', 'quantity'];
-const configTypesState = configTypes.reduce((types, name) => ({
-  ...types,
-  [`${name}Types`]: [],
-}), {});
-const requestConfigEntity = name =>
-  farm.remote.request(`/api/${name}_type/${name}_type`).then(prop('data'));
-const requestConfigDocs = () => Promise.all([
-  farm.schema.fetch(),
-  ...configTypes.map(requestConfigEntity),
-]).then(([schemata, ...configResponses]) => {
-  const configEntities = configResponses.reduce((ents, { data }, i) => ({
-    ...ents,
-    [configTypes[i]]: data,
-  }), {});
-  return [schemata, configEntities];
-});
-
 function parseFilterWithOptions(filter, options = {}) {
   const predicates = [parseFilter(filter)];
   if (options.includeUnsynced) predicates.push(farm.meta.isUnsynced);
@@ -78,7 +53,6 @@ export default {
     quantities: [],
     terms: [],
     users: [],
-    ...configTypesState,
   },
   mutations: {
     upsertEntity(state, { shortPlural, entity }) {
@@ -108,57 +82,8 @@ export default {
       state.terms = [];
       state.users = [];
     },
-    setConfigTypes(state, { name, types }) {
-      state[`${name}Types`] = types;
-    },
   },
   actions: {
-    loadConfigDocs({ commit }) {
-      const dbRequests = Object.keys(nomenclature.entities)
-        .map(name => getRecords('config_documents', name, d => d.enabled)
-          .then((docs) => {
-            const [schemata, configEntities] = partitionConfigDocs(docs);
-            farm.schema.set(name, schemata);
-            const types = map(pick(['label', 'name_pattern']), configEntities);
-            commit('setConfigTypes', { name, types });
-            return docs;
-          }));
-      return Promise.all(dbRequests);
-    },
-    updateConfigDocs({ commit }) {
-      return requestConfigDocs().then(([schemata, configEntities]) => {
-        farm.schema.set(schemata);
-        return databases.config_documents.stores.map(({ name }) => {
-          const docs = Object.entries(schemata[name] || {}).map(([key, json_schema]) => ({
-            key, json_schema, config_entity: null, enabled: true,
-          }));
-          if (name in configEntities) {
-            const types = {};
-            configEntities[name].forEach((config_entity) => {
-              const {
-                attributes: { drupal_internal__id: key, label, name_pattern },
-              } = config_entity;
-              types[key] = { label, name_pattern };
-              const doc = docs.find(d => d.key === key);
-              if (doc) {
-                doc.config_entity = config_entity;
-              } else {
-                docs.push({
-                  key, config_entity, json_schema: null, enabled: true,
-                });
-              }
-            });
-            commit('setConfigTypes', { name, types });
-          }
-          const enableRequest = Promise.all(docs.map(doc =>
-            saveRecord('config_documents', name, doc)));
-          const disableRequest = getRecords('config_documents', name, doc =>
-            none(d => d.key === doc.key, docs));
-          return Promise.all([enableRequest, disableRequest])
-            .then(([enabled, disabled]) => ({ name, enabled, disabled }));
-        });
-      });
-    },
     purgeEntities({ state }) {
       const now = Date.now();
       const uid = state.profile.user.id;
